APM-1358: Send MessagePack to the backend instead of JSON and Protobuf using the Python Agent

+++ .devops/tests.sh
@@ -7,8 +7,15 @@ if [ ! -d .auklet ]; then
     touch .auklet/version
 fi
 
+if [ -d htmlcov ]; then
+    rm -R htmlcov
+fi
+
 coverage run --rcfile=".coveragerc" setup.py test
-coverage html -d htmlcov
+coverage report -m
+coverage html -d tmp/htmlcov
 coverage xml
 
-rm -R .auklet
+if [ -d .auklet ]; then
+    rm -R .auklet
+fi
